+++ autoswitch_core/network.py
@@ -1,9 +1,8 @@
 """
 Networking Functions
 """
-
-from autoswitch_core.device import Device
 import os
+import netmiko
 import pprint
 
 
@@ -37,3 +36,28 @@ def load_with_serial(args: dict):
 
     with Device(**device_config) as device:
         device.load(args["load"])
+
+
+class Device:
+    def __init__(self, **kwargs) -> None:
+
+        kwargs.update({"default_enter": "\n",
+                       "global_delay_factor": 2})
+
+        self.conn = netmiko.ConnectHandler(**kwargs)
+
+        self.output = ""
+
+    def __enter__(self):
+        return self
+
+    def __exit__(self, type, value, traceback):
+        self.conn.disconnect()
+
+    def load(self, filename):
+        """
+        load config from txt file
+        """
+        self.conn.enable()
+        self.output = self.conn.send_config_from_file(filename)
+        print(self.output)
